refactor: 将 FmSetting.cs 拆分为 8 个 partial class 文件
FmSetting.cs — 仅保留字段声明
FmSetting.Init.cs — 构造函数/Form1_Load/readIniFile
FmSetting.AI.cs — AI OCR/翻译接口配置
FmSetting.Save.cs — 保存设置/关闭/路径检查
FmSetting.ApiKeys.cs — API密钥验证
FmSetting.General.cs — 常规设置/快捷键/代理/更新
FmSetting.OcrModel.cs — OCR模型路径配置
FmSetting.WebDav.cs — WebDAV备份恢复
FmSetting.Misc.cs — 杂项(音效/反馈/布局/自启动)
